+++ modules/dashboard/variables.tf
@@ -1,0 +1,8 @@
+variable "dashboard_name" {}
+variable "es_domain" {}
+variable "platform_instance_id" {}
+variable "proxy_asg" {}
+variable "rds_name" {}
+variable "region" {}
+variable "services_alb" {}
+variable "services_asg" {}