+++ modules/services/outputs.tf
@@ -9,3 +9,11 @@ output "services_security_group_id" {
 output "sns_topic_arn_harvest_complete" {
   value = "${aws_sns_topic.harvest_complete.arn}"
 }
+
+output "services_alb_cw" {
+  value = "${aws_lb.services_alb.arn_suffix}"
+}
+
+output "services_asg" {
+  value = "${aws_cloudformation_stack.services_asg.outputs["AsgName"]}"
+}